Remove derived wallet packages peer dependecies
