+++ .changeset/legal-bees-ring.md
@@ -1,0 +1,5 @@
+---
+"@aptos-labs/wallet-adapter-react": minor
+---
+
+Remove derived wallet packages peer dependecies
