+++ packages/wallet-adapter-react/package.json
@@ -42,18 +42,14 @@
     "@types/react-dom": "^18.3.0",
     "eslint": "^8.15.0",
     "tsup": "^8.4.0",
-    "typescript": "^5.4.5",
-    "@aptos-labs/derived-wallet-solana": "workspace:*",
-    "@aptos-labs/derived-wallet-ethereum": "workspace:*"
+    "typescript": "^5.4.5"
   },
   "dependencies": {
     "@aptos-labs/wallet-adapter-core": "workspace:*",
     "@radix-ui/react-slot": "^1.0.2"
   },
   "peerDependencies": {
-    "react": "^18.0.0 || ^19.0.0",
-    "@aptos-labs/derived-wallet-solana": "workspace:*",
-    "@aptos-labs/derived-wallet-ethereum": "workspace:*"
+    "react": "^18.0.0 || ^19.0.0"
   },
   "files": [
     "dist",
