+++ packages/wallet-adapter-react/src/useWallet.tsx
@@ -17,7 +17,6 @@ import {
   AptosSignInInput,
   AptosSignInOutput,
 } from "@aptos-labs/wallet-adapter-core";
-import { OriginWalletDetails } from "./WalletProvider";
 
 export interface WalletContextState {
   connected: boolean;
@@ -30,7 +29,7 @@ export interface WalletContextState {
     input: AptosSignInInput;
   }): Promise<AptosSignInOutput | void>;
   signAndSubmitTransaction(
-    transaction: InputTransactionData,
+    transaction: InputTransactionData
   ): Promise<AptosSignAndSubmitTransactionOutput>;
   signTransaction(args: {
     transactionOrPayload: AnyRawTransaction | InputTransactionData;
@@ -44,13 +43,8 @@ export interface WalletContextState {
   disconnect(): void;
   changeNetwork(network: Network): Promise<AptosChangeNetworkOutput>;
   submitTransaction(
-    transaction: InputSubmitTransactionData,
+    transaction: InputSubmitTransactionData
   ): Promise<PendingTransactionResponse>;
-  getOriginWalletDetails(
-    wallet: AdapterWallet,
-  ): Promise<OriginWalletDetails | undefined>;
-  isSolanaDerivedWallet(wallet: AdapterWallet): boolean;
-  isEIP1193DerivedWallet(wallet: AdapterWallet): boolean;
   wallet: AdapterWallet | null;
   wallets: ReadonlyArray<AdapterWallet>;
   notDetectedWallets: ReadonlyArray<AdapterNotDetectedWallet>;
@@ -61,7 +55,7 @@ const DEFAULT_CONTEXT = {
 };
 
 export const WalletContext = createContext<WalletContextState>(
-  DEFAULT_CONTEXT as WalletContextState,
+  DEFAULT_CONTEXT as WalletContextState
 );
 
 export function useWallet(): WalletContextState {
